Documentation and prepared restructuring.
Documentation:

Added a "docs/" folder and inserted the skeleton for mapper
documentation using the Documenter.jl package. Will expand
as I have time.

Restructuring:

Next step is to restructure the Mapper to remove the KiloCore
framework dependence. This means:

- Develop a minimal API that a KiloCore package can use to interact with
        the Mapper.
- Make a KiloCore package that uses the Mapper to achieve the current
        level of functionality.

Additionally, I would like to modularize parts of the Mapper that make
sense to modularize. That is, have an "Architecture" module, a
"Simulated Annealing" module, a "Routing" module etc. I'm not entirely
sure if this will make things better, but it will allow documentation
with Documenter.jl to be much cleaner.